2048 with basic solver and playable game
Use arrow keys to make moves and "p" to auto-move based on points and
"c" to auto-move based on complexity.  Uses a very basic solving
algorithm that just tests all the possible moves to a given depth and
responds with the best possible move, currently does not account for the
randomly added numbers after each move.